feat: "flatten" and functionalize API
Since inter-container communication is a bit of chore with Java, flatten
the architecture from Frontend->Servlet->Servlet->DB to
Frontend->Servlet+Controller->DB.

The controller part is the API module, now dependency of the two
frontends. Operations will be exposed as static classes in a rest-like
fashion: UserApi.xxx() will deal with Users etc. The API will comunicate
with the servlets using ApiResults, an Optional-like object part of the
evolving functional style infrastructure used by the api module.
ApiResult will carry a result or an ApiError, an easily json-alizable
object that will describe in detail the errors encountered.

The haskell-like infrastructure is still very rough and subject to
change. Points to be addressed in the near future:

1. Introduction of a try-catch-finally mechanism a bit more
   sophisticated than what Result.of() offers
2. Make UnavailableDatabase checked (or remove it) and make use more of
   Result
3. Cleanup the spaghetti of error-handling and code that is User