fix(static): show specific error message instead of a generic one
In `/details.html` and `/confirmTransfer.html`, if the corresponding
filter fails, the `error` attribute is passed in. Show the contained
message if said attribute is present. Only in case of a 500 code
show a beautified message since that can happen whenever and, unlike
the 4xx codes, is not result of a user tampering with the application.